# docs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Cloudflare Backend Boilerplate API",
		"version": "0.1.0",
		"description": "API-only Cloudflare Worker (Hono) for mobile clients. Generated from src/routes/* and src/lib/schemas.ts — see scripts/generate-openapi.ts."
	},
	"servers": [
		{
			"url": "http://127.0.0.1:8787",
			"description": "Local dev"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Better Auth JWT access token (15m). Also accepts an opaque session/API token on some routes — see docs/README."
			}
		}
	},
...
	"paths": {
...
		"/api/v1/api-keys": {
			"get": {
				"tags": [
					"API Keys"
				],
				"summary": "List the current user's API keys",
				"responses": {
					"200": {
						"description": "API keys (prefix only, no secret)",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"post": {
				"tags": [
					"API Keys"
				],
				"summary": "Create an API key",
				"responses": {
					"200": {
						"description": "The full secret — shown once",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string",
										"minLength": 1,
										"maxLength": 128
									}
								},
								"required": [
									"name"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/api-keys/{id}": {
			"delete": {
				"tags": [
					"API Keys"
				],
				"summary": "Revoke an API key",
				"responses": {
					"200": {
						"description": "API key revoked",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	}
}